attachments/ros_ws/requirements.txt
actionlib==1.12.1
angles==1.9.12
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
backports.functools-lru-cache==1.4
base-local-planner==1.16.7
bondpy==1.8.5
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
cbor==1.0.0
chardet==3.0.4
click==6.7
cloudpickle==1.2.2
colorama==0.3.7
configparser==4.0.2
constantly==15.1.0
controller-manager==0.18.2
controller-manager-msgs==0.18.2
cryptography==2.1.4
cv-bridge==1.13.0
cycler==0.10.0
Cython==0.21.1
defusedxml==0.5.0
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
docutils==0.14
dynamic-reconfigure==1.6.3
empy==3.3.2
enum34==1.1.10
future==0.18.2
futures==3.2.0
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.14
gym==0.16.0
hyperlink==17.3.1
idna==2.6
image-geometry==1.13.0
incremental==16.10.1
interactive-markers==1.11.5
ipaddress==1.0.17
joint-state-publisher==1.12.15
kdl-parser-py==1.13.1
keyring==10.6.0
keyrings.alt==3.0
laser-geometry==1.6.5
lz4==0.10.1
matplotlib==2.1.1
message-filters==1.14.10
mpi4py==2.0.0
netifaces==0.10.4
nose==1.3.7
numpy==1.16.6
olefile==0.45.1
PAM==0.4.2
paramiko==2.0.0
Pillow==6.2.2
psutil==5.4.2
py-ubjson==0.8.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.4.7
pydot==1.2.3
pyglet==1.5.0
Pygments==2.2.0
pygobject==3.26.1
PyNaCl==1.1.2
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyrvo2==0.0.0
pyserial==3.4
python-dateutil==2.6.1
python-gnupg==0.4.1
python-qt-binding==0.4.3
python-snappy==0.5
PyTrie==0.2
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
qrcode==5.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource-retriever==1.12.6
roman==2.0.0
rosbag==1.14.10
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosdep-modules==0.19.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.14.10
roslaunch==1.14.10
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.10
rosmake==1.14.9
rosmaster==1.14.10
rosmsg==1.14.10
rosnode==1.14.10
rosparam==1.14.10
rospkg==1.2.8
rospkg-modules==1.2.8
rospy==1.14.10
rosservice==1.14.10
rostest==1.14.10
rostopic==1.14.10
rosunit==1.14.9
roswtf==1.14.10
rqt-action==0.4.9
rqt-bag==0.4.14
rqt-bag-plugins==0.4.14
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.13.14
scipy==1.2.3
SecretStorage==2.3.1
sensor-msgs==1.12.7
service-identity==16.0.0
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
subprocess32==3.2.7
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.10
torch==1.4.0
torchvision==0.5.0
trollius==2.0.1
Twisted==17.9.0
txaio==2.8.1
u-msgpack-python==2.1
urdfdom-py==0.4.4
wsaccel==0.6.2
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xacro==1.13.9
zope.interface==4.3.2
